+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-19 03:03:58
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 35658e1f358eb91fd07a0581641b802bc96d9546
+
+## [ForwardWidgets-2kuai] - 2025-08-19 03:03:58
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: d685d2a56ccb48b0cb04048b02c952a4c50dd9c0
+
+## [ForwardWidgets-pack1r] - 2025-08-19 03:03:58
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 74ff483217877b00467f65edb2d53f3ee78197ab
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-18 03:22:00
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main